update-requirements.txt
pandas
tabula-py
plotly
openpyxl
pymsteams
python-dotenv
kaleido
deta